# apps/api/src/modules/provider-work/provider-work.controller.ts
import type { FastifyReply } from 'fastify';
import type { FastifyPluginAsyncZod } from 'fastify-type-provider-zod';
import { z } from 'zod';
import {
  CreatePromptGenerationBodySchema,
  CompletePromptGenerationBodySchema,
  ClaimWorkRequestBodySchema,
  CreateIvrAiTurnBodySchema,
  CompleteIvrAiTurnBodySchema,
} from '@managecallai/contracts';
import { db } from '../../db/client.js';
import type { AuthClaims } from '../auth/auth-claims.js';
import { CAPABILITIES } from '../auth/capabilities.js';
import { requireCapability } from '../auth/require-capability.js';
import { authenticateRuntime } from '../runtime/runtime-auth.js';
import { ProviderWorkRepository } from './provider-work.repository.js';
import { ProviderWorkRequestNotFoundError, ProviderWorkService } from './provider-work.service.js';
import { sendNotFound, sendInvalidArgument } from '../../errors/index.js';

const service = new ProviderWorkService(new ProviderWorkRepository(db));

function replyNotFound(err: unknown, reply: FastifyReply): void {
  if (err instanceof ProviderWorkRequestNotFoundError) {
    return sendNotFound(reply, err.message);
  }
  throw err;
}

export const promptGenerationController: FastifyPluginAsyncZod = async (app) => {
  app.get(
    '/requests',
    { preHandler: requireCapability(CAPABILITIES.TENANT_PROMPTS_VIEW) },
    async (req) => {
      const user = req.user as AuthClaims;
      return { data: await service.listPromptGenerations(user.tenant_id) };
    },
  );

  app.post(
    '/requests',
    {
      preHandler: requireCapability(CAPABILITIES.TENANT_PROMPTS_CREATE),
      schema: {
        body: CreatePromptGenerationBodySchema,
      },
    },
    async (req, reply) => {
      const user = req.user as AuthClaims;
      const request = await service.createPromptGeneration(user.tenant_id, req.body);
      return reply.code(201).send({ data: request });
    },
  );

  app.get(
    '/requests/:requestId',
    {
      preHandler: requireCapability(CAPABILITIES.TENANT_PROMPTS_VIEW),
      schema: {
        params: z.object({ requestId: z.string() }),
      },
    },
    async (req, reply) => {
      const user = req.user as AuthClaims;
      try {
        return { data: await service.getPromptGeneration(req.params.requestId, user.tenant_id) };
      } catch (err) {
        return replyNotFound(err, reply);
      }
    },
  );

  app.post(
    '/internal/:requestId/claim',
    {
      preHandler: authenticateRuntime,
      schema: {
        params: z.object({ requestId: z.string() }),
        body: ClaimWorkRequestBodySchema,
      },
    },
    async (req, reply) => {
      try {
        return { data: await service.claimPromptGeneration(req.params.requestId, req.body) };
      } catch (err) {
        return replyNotFound(err, reply);
      }
    },
  );

  app.post(
    '/internal/:requestId/result',
    {
      preHandler: authenticateRuntime,
      schema: {
        params: z.object({ requestId: z.string() }),
        body: CompletePromptGenerationBodySchema,
      },
    },
    async (req, reply) => {
      try {
        return { data: await service.completePromptGeneration(req.params.requestId, req.body) };
      } catch (err) {
        return replyNotFound(err, reply);
      }
    },
  );
};

export const ivrAiController: FastifyPluginAsyncZod = async (app) => {
  app.post(
    '/runtime/ivr-ai/turns',
    {
      preHandler: authenticateRuntime,
      schema: {
        body: CreateIvrAiTurnBodySchema.extend({ tenant_id: z.string().min(1).optional() }),
      },
    },
    async (req, reply) => {
      const tenantId = req.body.tenant_id ?? '';
      if (!tenantId) return sendInvalidArgument(reply, 'tenant_id is required for runtime IVR AI turns');
      const request = await service.createIvrAiTurn(tenantId, req.body);
      return reply.code(201).send({ data: request });
    },
  );

  app.get(
    '/runtime/ivr-ai/turns/:requestId',
    {
      preHandler: requireCapability(CAPABILITIES.TENANT_IVR_FLOWS_VIEW),
      schema: {
        params: z.object({ requestId: z.string() }),
      },
    },
    async (req, reply) => {
      const user = req.user as AuthClaims;
      try {
        return { data: await service.getIvrAiTurn(req.params.requestId, user.tenant_id) };
      } catch (err) {
        return replyNotFound(err, reply);
      }
    },
  );

  app.post(
    '/ivr-ai/internal/:requestId/claim',
    {
      preHandler: authenticateRuntime,
      schema: {
        params: z.object({ requestId: z.string() }),
        body: ClaimWorkRequestBodySchema,
      },
    },
    async (req, reply) => {
      try {
        return { data: await service.claimIvrAiTurn(req.params.requestId, req.body) };
      } catch (err) {
        return replyNotFound(err, reply);
      }
    },
  );

  app.post(
    '/ivr-ai/internal/:requestId/result',
    {
      preHandler: authenticateRuntime,
      schema: {
        params: z.object({ requestId: z.string() }),
        body: CompleteIvrAiTurnBodySchema,
      },
    },
    async (req, reply) => {
      try {
        return { data: await service.completeIvrAiTurn(req.params.requestId, req.body) };
      } catch (err) {
        return replyNotFound(err, reply);
      }
    },
  );
};