# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: manageCallAI API
  version: 0.1.0
  summary: Business-level telecom control plane API over FreeSWITCH
  description: |
    OpenAPI contract for the manageCallAI REST API.

    **MVP response envelope**: all implemented endpoints return `{ "data": ... }` for
    both single-resource and list responses. Pagination will be introduced later
    where required. Error responses return `{ "error": "<CODE>", "message": "<human text>",
    "request_id": "<id>" }` where `error` is a machine-readable RPC code (NOT_FOUND,
    INVALID_ARGUMENT, UNAUTHENTICATED, PERMISSION_DENIED, ALREADY_EXISTS,
    RESOURCE_EXHAUSTED, INTERNAL, UNAVAILABLE).

    **Auth model**:
    - Human-facing endpoints: `Authorization: Bearer <JWT>` (issued by `/auth/login`
      or `/auth/register`). JWT claims include `sub`, `tenant_id`, `email`, `role`.
    - Runtime/internal endpoints: `Authorization: Bearer <RUNTIME_API_TOKEN>` or
      header `x-managecallai-runtime-token: <token>`. These are called by
      FreeSWITCH and the Go ESL agent — not by end users.

    Only implemented endpoints are documented here. Future resources (outbound
    routes, prompts, CDRs, audit events, approvals, simulations) will be added
    when implemented.
servers:
  - url: https://api.example.com/api/v1
    description: Production placeholder
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /phone-numbers:
    get:
      tags:
        - PhoneNumbers
      summary: List phone numbers
      responses:
        "200":
          description: Phone number list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumberListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags:
        - PhoneNumbers
      summary: Create phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePhoneNumberRequest"
      responses:
        "201":
          description: Phone number created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumberResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /phone-numbers/{numberId}:
    parameters:
      - $ref: "#/components/parameters/NumberIdParam"
    get:
      tags:
        - PhoneNumbers
      summary: Get phone number
      responses:
        "200":
          description: Phone number
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumberResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags:
        - PhoneNumbers
      summary: Update phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePhoneNumberRequest"
      responses:
        "200":
          description: Phone number updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumberResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /phone-numbers/{numberId}/deactivate:
    parameters:
      - $ref: "#/components/parameters/NumberIdParam"
    post:
      tags:
        - PhoneNumbers
      summary: Deactivate phone number
      responses:
        "200":
          description: Phone number deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PhoneNumberResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /prompts:
    get:
      tags:
        - Prompts
      summary: List prompt assets
      responses:
        "200":
          description: Prompt asset list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptAssetListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags:
        - Prompts
      summary: Create prompt asset metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromptAssetRequest"
      responses:
        "201":
          description: Prompt asset created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptAssetResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /prompts/{promptId}:
    parameters:
      - $ref: "#/components/parameters/PromptIdParam"
    get:
      tags:
        - Prompts
      summary: Get prompt asset
      responses:
        "200":
          description: Prompt asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptAssetResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags:
        - Prompts
      summary: Update prompt asset metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePromptAssetRequest"
      responses:
        "200":
          description: Prompt asset updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptAssetResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /prompts/{promptId}/deactivate:
    parameters:
      - $ref: "#/components/parameters/PromptIdParam"
    post:
      tags:
        - Prompts
      summary: Deactivate prompt asset
      responses:
        "200":
          description: Prompt asset deactivated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptAssetResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /runtime/ivr-ai/turns:
    post:
      tags:
        - IvrAI
      summary: Create an AI-assisted IVR turn request
      description: |
        Runtime-internal contract for IVR nodes that ask an external AI service to
        answer a caller question, classify intent, or generate a spoken/text response.
        Provider execution is external to the core platform. OpenAI-compatible,
        custom, or tenant-installed providers can implement this contract.
      security:
        - runtimeTokenHeader: []
        - runtimeTokenBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIvrAiTurnRequest"
      responses:
        "201":
          description: IVR AI turn request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrAiTurnRequestResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /runtime/ivr-ai/turns/{ivrAiTurnRequestId}:
    get:
      tags:
        - IvrAI
      summary: Get IVR AI turn request status and result
      description: Tenant-facing status read for a pending or completed AI turn.
      parameters:
        - $ref: "#/components/parameters/IvrAiTurnRequestIdParam"
      responses:
        "200":
          description: IVR AI turn request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrAiTurnRequestResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-ai/internal/{ivrAiTurnRequestId}/claim:
    post:
      tags:
        - IvrAI
      summary: Claim an IVR AI turn request for processing
      security:
        - runtimeTokenHeader: []
        - runtimeTokenBearer: []
      parameters:
        - $ref: "#/components/parameters/IvrAiTurnRequestIdParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimIvrAiTurnRequest"
      responses:
        "200":
          description: IVR AI turn request claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrAiTurnRequestResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /ivr-ai/internal/{ivrAiTurnRequestId}/result:
    post:
      tags:
        - IvrAI
      summary: Submit IVR AI turn result from a trusted processor
      description: |
        Trusted callback for AI-generated caller response, intent classification, or
        selected next action. Public tenant users cannot call this endpoint.
      security:
        - runtimeTokenHeader: []
        - runtimeTokenBearer: []
      parameters:
        - $ref: "#/components/parameters/IvrAiTurnRequestIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteIvrAiTurnRequest"
      responses:
        "200":
          description: IVR AI turn result accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IvrAiTurnRequestResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT issued by /auth/login or /auth/register
    runtimeTokenBearer:
      type: http
      scheme: bearer
      description: RUNTIME_API_TOKEN as Bearer token (preferred)
    runtimeTokenHeader:
      type: apiKey
      in: header
      name: x-managecallai-runtime-token
      description: RUNTIME_API_TOKEN as custom header
  parameters:
...
    NumberIdParam:
      name: numberId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PromptIdParam:
      name: promptId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IvrAiTurnRequestIdParam:
      name: ivrAiTurnRequestId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          enum:
            - NOT_FOUND
            - INVALID_ARGUMENT
            - UNAUTHENTICATED
            - PERMISSION_DENIED
            - ALREADY_EXISTS
            - RESOURCE_EXHAUSTED
            - INTERNAL
            - UNAVAILABLE
        message:
          type: string
        request_id:
          type: string
      required:
        - error
        - message
        - request_id
...